Refactored class GenericError
